+++ src/app/login/page.tsx
@@ -1,7 +1,7 @@
 import { type Metadata } from "next";
 import Image from "next/image";
 
-import { LoginForm } from "@/components/forms/LoginForm";
+import { LoginForm } from "@/components/forms/login-form";
 import { PageHeading } from "@/components/ui/page-heading";
 
 export const metadata: Metadata = {
